+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [v1.8.3] - 2023-07-12
+
 ### Added
 
 - Allow to pass a Dict with field names and summary strategies to the `fields` parameter in the `Summarizer` constructor ([#1195](https://github.com/stac-utils/pystac/pull/1195))
@@ -762,7 +764,8 @@ use `Band.create`
 
 Initial release.
 
-[Unreleased]: <https://github.com/stac-utils/pystac/compare/v1.8.2..main>
+[Unreleased]: <https://github.com/stac-utils/pystac/compare/v1.8.3..main>
+[v1.8.3]: <https://github.com/stac-utils/pystac/compare/v1.8.2..v1.8.3>
 [v1.8.2]: <https://github.com/stac-utils/pystac/compare/v1.8.1..v1.8.2>
 [v1.8.1]: <https://github.com/stac-utils/pystac/compare/v1.8.0..v1.8.1>
 [v1.8.0]: <https://github.com/stac-utils/pystac/compare/v1.7.3..v1.8.0>